Use SSHBase for SFTP and SCP
