+++ lib/backup/storage/scp.rb
@@ -3,34 +3,14 @@ require 'net/scp'
 
 module Backup
   module Storage
-    class SCP < Base
+    class SCP < SSHBase
       include Storage::Cycler
-      class Error < Backup::Error; end
-
-      ##
-      # Server credentials
-      attr_accessor :username, :password, :ssh_options
-
-      ##
-      # Server IP Address and SCP port
-      attr_accessor :ip, :port
 
       def initialize(model, storage_id = nil)
         super
-
-        @port ||= 22
-        @path ||= 'backups'
-        @ssh_options ||= {}
-        path.sub!(/^~\//, '')
       end
 
       private
-
-      def connection
-        Net::SSH.start(
-          ip, username, { :password => password, :port => port }.merge(ssh_options)
-        ) {|ssh| yield ssh }
-      end
 
       def transfer!
         connection do |ssh|
